requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
gunicorn==20.1.0
azure-storage-blob==12.9.0
azure-storage-queue==12.1.6
python-multipart==0.0.5
pandas==1.3.4